Adds generic sensors and generic observers
Uses SensorStateMachines and SystemStateMachines to hold states and state transition logic. A StateManager is notified of any sensor driven state changes and tells the system dependent state machines to reassess their state after checking the relative upstream state machines.

+++ src/sensors/specific_sensors/test_i2c_sensor.cpp
@@ -1,5 +1,4 @@
 #include <src/sensors/specific_sensors/test_i2c_sensor.h>
-#include <src/observers/observer.h>
 
 TestI2cSensor::TestI2cSensor(std::string id) : id(id) {}
 
@@ -10,6 +9,6 @@ void TestI2cSensor::TakeReading() {
     new_reading = false;
 }
 
-void TestI2cSensor::SetDummySensorData(int dummy_data) {
+void TestI2cSensor::SetDummySensorData(double dummy_data) {
     this->dummy_data = dummy_data;
 }
